src/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
black==19.10b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-crum==0.7.9
django-rest-framework==0.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
flake8==3.8.4
flake8-broken-line==0.3.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
future==0.18.2
idna==3.2
importlib-metadata==4.5.0
inflection==0.5.1
isort==5.9.1
itypes==1.2.0
Jinja2==3.0.1
mando==0.6.4
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
pathspec==0.8.1
pep8-naming==0.11.1
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytz==2021.1
radon==4.3.2
regex==2021.4.4
requests==2.26.0
ruamel.yaml==0.17.14
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.4.1
